third_party/WebKit/LayoutTests/media/track/track-webvtt-tc022-entities.html

<!DOCTYPE html>
<title>Tests entities in the cue text.</title>
<script src="track-helpers.js"></script>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>
// TODO(srirama.m): Rewrite the test in a better way.
// See https://codereview.chromium.org/2030383002/ for details.
check_cues_from_track("captions-webvtt/tc022-entities.vtt", function(track) {
    var expected = [
        { innerHTML: "This cue has an ampersand & character." },
        { innerHTML: "This cue has a less than < character." },
        { innerHTML: "This cue has a greater than > character." },
        { innerHTML: "This cue has a Left-to-Right Mark \u200e." },
        { innerHTML: "This cue has a Right-to-Left Mark \u200f." },
        { innerHTML: "This cue has a non-breaking space \u00a0." },
        { innerHTML: "This & is parsed to the same as &." }
    ];

    var cues = track.cues;
    assert_equals(cues.length, expected.length);
    for (var i = 0; i < cues.length; i++)
        assert_equals(cues[i].getCueAsHTML().textContent, expected[i].innerHTML);
});

check_cues_from_track("captions-webvtt/tc022-entities-wrong.vtt", function(track) {
    var expected = [
        { innerHTML: "This cue has a less than ", },
        { innerHTML: "This cue has a greater than > character.\nSince it's not related to a < character,\nit's just interpreted as text.", }
    ];

    var cues = track.cues;
    assert_equals(cues.length, expected.length);
    for (var i = 0; i < cues.length; i++)
        assert_equals(cues[i].getCueAsHTML().textContent, expected[i].innerHTML);
});
</script>
